now keeps custom maps that are not in the saves directory

+++ src/main/java/modmuss50/HardCoreMapRest/ResetMaps.java
@@ -10,21 +10,38 @@ import java.io.*;
 public class ResetMaps {
 
     public static void resetMaps(){
-        Minecraft mc = Minecraft.getMinecraft();
-        File saveDir = new File(mc.mcDataDir, "saves");
-        File backupDir = new File(mc.mcDataDir, "maps");
-        deleteFolder(saveDir);
-        saveDir.mkdir();
-        try {
-            copyDirectory(backupDir, saveDir);
-        } catch (IOException e) {
-            e.printStackTrace();
+        if(!MapRest.keepOldMaps){
+            Minecraft mc = Minecraft.getMinecraft();
+            File saveDir = new File(mc.mcDataDir, "saves");
+            File backupDir = new File(mc.mcDataDir, "maps");
+            deleteFolder(saveDir);
+            saveDir.mkdir();
+            try {
+                copyDirectory(backupDir, saveDir);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            Minecraft.getMinecraft().displayGuiScreen(new GuiSelectWorld(new GuiMainMenu()));
+        } else {
+            Minecraft mc = Minecraft.getMinecraft();
+            File saveDir = new File(mc.mcDataDir, "saves");
+            File backupDir = new File(mc.mcDataDir, "maps");
+            String[] savechildren = saveDir.list();
+            String[] backchildren = backupDir.list();
+            for (int i = 0; i < savechildren.length; i++) {
+                for (int j = 0; j < backchildren.length; j++) {
+                    if(savechildren[i].equals(backchildren[j])){
+                        System.out.println("found a map to restore" + savechildren[i] + ":" + backchildren[j]);
+                        resetmap(savechildren[i]);
+                    }
+                }
+            }
+
+            Minecraft.getMinecraft().displayGuiScreen(new GuiSelectWorld(new GuiMainMenu()));
         }
-
-
-        Minecraft.getMinecraft().displayGuiScreen(new GuiSelectWorld(new GuiMainMenu()));
-
     }
+
+
 
     public static void copyDirectory(File sourceLocation , File targetLocation) throws IOException {
         if (sourceLocation.isDirectory()) {
@@ -66,5 +83,22 @@ public class ResetMaps {
         }
         folder.delete();
     }
+
+    public static void resetmap(String name){
+        Minecraft mc = Minecraft.getMinecraft();
+        File saveDir = new File(mc.mcDataDir, "saves");
+        File backupDir = new File(mc.mcDataDir, "maps");
+        File oldDir = new File(saveDir, name);
+        File newDir = new File(backupDir, name);
+        deleteFolder(oldDir);
+        oldDir.mkdir();
+        try {
+            copyDirectory(newDir, oldDir);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+
 
 }
